refactored RecursiveResolver and UriResolverAggregator from traits
