+++ packages/resolvers/src/resolver_with_loop_guard.rs
