+++ packages/resolvers/src/uri_resolver_aggregator.rs
@@ -15,7 +15,7 @@ pub struct UriResolverAggregator {
 }
 
 impl UriResolverAggregator {
-    pub fn new(resolvers: Vec<Box<dyn UriResolver>>) -> Self {
+    pub fn new(resolvers: Vec<Arc<dyn UriResolver>>) -> Self {
         let resolvers = resolvers.into_iter().map(Arc::from).collect();
         Self {
             name: None,
@@ -81,13 +81,14 @@ impl fmt::Debug for UriResolverAggregator {
   }
 }
 
-// impl From<Vec<Arc<dyn UriResolver>>> for UriResolverAggregator {
-//     fn from(resolvers: Vec<Arc<dyn UriResolver>>) -> Self {
-//         UriResolverAggregator::new(resolvers)
-//     }
-// }
 impl From<Vec<Box<dyn UriResolver>>> for UriResolverAggregator {
     fn from(resolvers: Vec<Box<dyn UriResolver>>) -> Self {
+        UriResolverAggregator::new(resolvers.into_iter().map(Arc::from).collect())
+    }
+}
+
+impl From<Vec<Arc<dyn UriResolver>>> for UriResolverAggregator {
+    fn from(resolvers: Vec<Arc<dyn UriResolver>>) -> Self {
         UriResolverAggregator::new(resolvers)
     }
 }
